refactor(StatusItemController): extract StatusItemImageUpdater
Moves the menu-bar strip render loop — updateButtonImage /
scheduleButtonImageUpdate / renderButtonImage + the pendingRenderTask debounce
+ the withObservationTracking re-arm — into a focused StatusItemImageUpdater.
It needs only the container (layout + dataStore) and a closure to set the
button image, so the controller is left owning panel construction, the status
item + context menu, transparency, the outside-click monitors, and show/hide.

The render/debounce/re-arm logic is relocated byte-for-byte (same
withObservationTracking pattern, which is self-contained), so behavior is
unchanged. StatusItemController: 631 → 593 LOC.

Verified: swift build + 883 tests green; release build launches, "Status item
ready (button: true)" — the strip renders on init, providers refresh, no crash.
